feat: it reduces the complexity of auth provider
