+++ src/constants/index.ts
@@ -1,3 +1,4 @@
 export * from './auth'
 export * from './common'
+export * from './create-medicine'
 export * from './medications'
