+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './getMedicationsParamsReducer'
+export * from './useAuth'
 export * from './useCreateMedicine'
 export * from './useQueryParams'
