+++ src/hooks/useAuth.ts
@@ -1,0 +1,13 @@
+import { useContext } from 'react'
+
+import { AuthContext } from '@/contexts'
+
+export const useAuth = () => {
+  const authContext = useContext(AuthContext)
+
+  if (!authContext) {
+    throw new Error('useAuth is not defined.')
+  }
+
+  return authContext
+}
