+++ src/providers/index.ts
@@ -1,2 +1,3 @@
 export * from './appProvider'
+export * from './authProvider'
 export * from './queryProvider'
