+++ src/utils/functions/handleFormsSubmitError.ts
@@ -5,12 +5,11 @@ import { ValidationError } from 'yup'
 export const handleFormsSubmitError = (
   error: unknown,
   setErrors: (errors: FormikErrors<unknown>) => void,
+  message: string,
 ) => {
   if (error instanceof ValidationError) {
     setErrors({ [error.path as string]: error.message })
     return
   }
-  toast.error(
-    'It was not possible to create the medication. Try again or contact support.',
-  )
+  toast.error(message)
 }
